serialisation.cr: Remove typeof(), fixes #2
The `typeof()` was a left-over of the initial work.  The macro already
expects a `type`, so `typeof(type)` gets us the type of `type`, which
in turn is passed into `sizeof()` - Thus we're getting the size of the
class type, and not of a variable of the type.  Crystal 0.23 fixed this
errornous behaviour, and thus broke Cannon - For good reason.